doc: Sanity check for building docs from the right directory

+++ bin/compile_doc.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if ! [ -e README.org ]; then
+    echo >&2 "Error: Please run this script from top of the Organice repository."
+    exit 1
+fi
+
 echo "#+SETUPFILE: https://fniessen.github.io/org-html-themes/setup/theme-readtheorg.setup" > documentation.org
 cat README.org | grep -v api.codeclimate | grep -v "https://organice.200ok.ch/documentation.html" >> documentation.org
 sed -i 's/# REPO_PLACEHOLDER/Code repository: https:\/\/github.com\/200ok-ch\/organice/' documentation.org
